DTO created for Course, API endpoints updated

+++ src/main/java/com/davidwerfelmann/course_scheduler/data/CourseRepository.java
@@ -3,5 +3,5 @@ package com.davidwerfelmann.course_scheduler.data;
 import com.davidwerfelmann.course_scheduler.models.Course;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CourseRepository extends CrudRepository<Course, Integer> {
+public interface CourseRepository extends CrudRepository<Course, Long> {
 }
